my_sample_dbt/requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.17.1
dbt-common==1.32.0
dbt-core==1.10.13
dbt-extractor==0.6.0
dbt-postgres==1.9.1
dbt-protos==1.0.375
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
idna==3.10
importlib_metadata==8.7.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
leather==0.4.0
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.1
networkx==3.5
orderly-set==5.5.0
packaging==25.0
parsedatetime==2.6
pathspec==0.12.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
snowplow-tracker==1.1.0
sqlparse==0.5.3
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
